android: ion: mtk: fix clang -Wvoid-pointer-to-int-cast
(cherry picked from commit c5970a9d14dc83a2c144f9f1a0340f563622c672)
(cherry picked from commit 45957c940d037ec0095e74bdcabdbf4270dfa025)
(cherry picked from commit 26d1672d705bc485176e3403e1aff511042fbd39)

+++ drivers/staging/android/ion/ion.c
@@ -122,7 +122,7 @@ void ion_client_buf_add(struct ion_heap *heap, struct ion_client *client,
 	u64 total_size;
 
 	client->hnd_cnt++;
-	if (heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC)
+	if ((int)heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC)
 		total_size =
 		atomic64_add_return(size, &client->total_size[SECURE_HEAP]);
 	else if (heap->type == ION_HEAP_TYPE_SYSTEM)
@@ -162,7 +162,7 @@ void ion_client_buf_sub(struct ion_heap *heap, struct ion_client *client,
 	long long total_size;
 
 	client->hnd_cnt--;
-	if (heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC) {
+	if ((int)heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC) {
 		total_size =
 		atomic64_sub_return(size, &client->total_size[SECURE_HEAP]);
 		if (total_size < 0) {
@@ -207,7 +207,7 @@ void ion_client_buf_sub(struct ion_heap *heap, struct ion_client *client,
 u64 ion_client_buf_dump(struct ion_heap *heap, struct ion_client *client)
 {
 #ifdef ION_RECORD_TOTAL_SIZE_SUPPORT
-	if (heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC)
+	if ((int)heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC)
 		return (u64)(atomic64_read(&client->total_size[SECURE_HEAP]));
 	else if (heap->type == ION_HEAP_TYPE_SYSTEM)
 		return (u64)(atomic64_read(&client->total_size[SYSTEM_HEAP]));
@@ -360,7 +360,7 @@ static struct ion_buffer *ion_buffer_create(struct ion_heap *heap,
 	INIT_LIST_HEAD(&buffer->attachments);
 #endif
 	mutex_init(&buffer->lock);
-	if (heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC)
+	if ((int)heap->type == ION_HEAP_TYPE_MULTIMEDIA_SEC)
 		goto exit;
 	/*
 	 * this will set up dma addresses for the sglist -- it is not
@@ -2203,8 +2203,8 @@ static size_t ion_debug_heap_total(struct ion_client *client,
 			/* for exception dump ion_mm_heap info */
 		    (id == ION_HEAP_TYPE_MULTIMEDIA &&
 		     (type == ION_HEAP_TYPE_SYSTEM ||
-		      type == ION_HEAP_TYPE_MULTIMEDIA ||
-		      type == ION_HEAP_TYPE_MULTIMEDIA_SEC))) {
+		      (int)type == ION_HEAP_TYPE_MULTIMEDIA ||
+		      (int)type == ION_HEAP_TYPE_MULTIMEDIA_SEC))) {
 			client->dbg_hnd_cnt++;
 			size += handle->buffer->size;
 		}
